Kernel with working interrupts :D

+++ cpu/timer.h
@@ -8,7 +8,7 @@ void init_timer(uint32_t freq);
 void wait(uint32_t ms);
 void sleep(uint32_t ms);
 
-uint32_t tick;
+uint64_t tick;
 int lSnd;
 int pSnd;
 int task;
